break(router): move `IncomingCloudflareProperties` to `worktop/cfw`
