+++ packages/worktop/types/cfw.ts
@@ -1,6 +1,6 @@
 import type { Module } from 'worktop/cfw';
 import type { Bindings, CronEvent } from 'worktop';
-import type { IncomingCloudflareProperties } from 'worktop';
+import type { IncomingCloudflareProperties } from 'worktop/cfw';
 import type { Strict } from 'worktop/utils';
 import type { KV } from 'worktop/kv';
